added ability to load Rayforce geometry from a passed Mesh via setGroup()
-Previously, the only way to get geometry into the Rayforce AS was to
load a graph cache file. Using the mesh we get from setGroup() makes sure
that we get all the data we can from the other parts of Manta, thus being
a more robust route.

-Also added the ability to save out the graph cache via the Rayforce AS.
This will thus be loaded exactly into memory as it was before. It has not
been tested, though...